Set up routing for app
